Transferring code from Steps to VolunteerBrief
